add model api
