+++ setup.cfg
@@ -1,0 +1,8 @@
+[metadata]
+name = mypackage
+version = 0.0.1
+
+[options]
+install_requires =
+    requests
+    importlib-metadata; python_version < "3.8"
